+++ src/WifiConfigParameters.h
@@ -15,6 +15,10 @@ public:
     WiFiManagerParameter &tempOffsetIndoor();
     WiFiManagerParameter &outdoorSensorChannel();
     WiFiManagerParameter &ventingThreshold();
+    WiFiManagerParameter &mqttHost();
+    WiFiManagerParameter &mqttPort();
+    WiFiManagerParameter &mqttUsername();
+    WiFiManagerParameter &mqttPassword();
 
 private:
     String timezoneSelectHtml;
@@ -22,6 +26,7 @@ private:
     String tempOffsetIndoorValue;
     String outdoorSensorChannelValue;
     String ventingThresholdValue;
+    String mqttPortValue;
 
     WiFiManagerParameter timezoneSelectParam;
     WiFiManagerParameter languageSelectParam;
@@ -31,7 +36,12 @@ private:
     WiFiManagerParameter tempOffsetIndoorParam;
     WiFiManagerParameter outdoorSensorChannelParam;
     WiFiManagerParameter ventingThresholdParam;
+    WiFiManagerParameter mqttHostParam;
+    WiFiManagerParameter mqttPortParam;
+    WiFiManagerParameter mqttUsernameParam;
+    WiFiManagerParameter mqttPasswordParam;
 
     static String formatFloatValue(float value, uint8_t decimals);
     static String formatIntegerValue(uint8_t value);
+    static String formatMqttPortValue(uint16_t value);
 };